requirements.txt
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
defusedxml==0.7.1
Django==4.2.2
django-cors-headers==4.1.0
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
httpie==3.2.2
idna==3.4
importlib-resources==5.12.0
inflection==0.5.1
install==1.3.5
jsonschema==4.17.3
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
pkgutil_resolve_name==1.3.10
Pygments==2.15.1
PyJWT==2.7.0
pyrsistent==0.19.3
PySocks==1.7.1
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.4.2
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
zipp==3.15.0
